NUP-2386 Renamed RecordSensor.py to record_sensor.py, but didn't update the builtin regions registration to demonstrate failure witnessed by Matt
