+++ src/nupic/data/stats.py
@@ -24,7 +24,7 @@ import pickle
 
 from pkg_resources import resource_filename
 
-from nupic.regions.RecordSensor import RecordSensor
+from nupic.regions.record_sensor import RecordSensor
 from nupic.data.file_record_stream import FileRecordStream
