+++ src/nupic/regions/record_sensor.py
