+++ tests/integration/nupic/engine/network_checkpoint_test.py
@@ -23,7 +23,7 @@ import unittest
 import numpy
 
 from nupic.regions.SPRegion import SPRegion
-from nupic.regions.RecordSensor import RecordSensor
+from nupic.regions.record_sensor import RecordSensor
 
 from network_creation_common import createAndRunNetwork
