+++ tests/integration/nupic/engine/network_creation_common.py
@@ -29,7 +29,7 @@ from pkg_resources import resource_filename
 from nupic.data.file_record_stream import FileRecordStream
 from nupic.engine import Network
 from nupic.encoders import MultiEncoder, ScalarEncoder, DateEncoder
-from nupic.regions.RecordSensor import RecordSensor
+from nupic.regions.record_sensor import RecordSensor
 from nupic.regions.SPRegion import SPRegion
 from nupic.regions.TMRegion import TMRegion
